+++ src/constants/settings.constants.ts
@@ -1,4 +1,5 @@
 import { DEFAULT_WHEEL_ACTION } from '@/constants/camera.constants';
+import { DEFAULT_KEYMAP } from '@/constants/keymap.constants';
 import {
   DEFAULT_COLOR,
   DEFAULT_ERASER_RADIUS,
@@ -18,6 +19,7 @@ export const DEFAULT_SETTINGS: Settings = {
   penSize: DEFAULT_SIZE,
   eraserRadius: DEFAULT_ERASER_RADIUS,
   wheelAction: DEFAULT_WHEEL_ACTION,
+  keymap: DEFAULT_KEYMAP,
   sortOrder: 'modified',
   lastSeenVersion: '',
 };