+++ src/renderer/board/tools/pen.tool.ts
@@ -2,6 +2,6 @@ import { StrokeTool } from '@/renderer/board/tools/stroke.tool';
 
 export class PenTool extends StrokeTool {
   constructor() {
-    super('pen', 'Pen', ['p', '1'], 'pen');
+    super('pen', 'Pen', 'pen');
   }
 }