+++ src/renderer/board/tools/pencil.tool.ts
@@ -2,6 +2,6 @@ import { StrokeTool } from '@/renderer/board/tools/stroke.tool';
 
 export class PencilTool extends StrokeTool {
   constructor() {
-    super('pencil', 'Pencil', ['n', '2'], 'pencil');
+    super('pencil', 'Pencil', 'pencil');
   }
 }